refactor : S3 Service 레이어가 S3 라이브러리를 의존하고 있어서 이를 분리
